backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
requests
PyYAML  
python-dotenv